requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
requests
google-cloud-aiplatform
google-generativeai
google-genai